+++ src/main/kotlin/org/neo4j/graphql/handler/CreateTypeHandler.kt
@@ -5,14 +5,12 @@ import graphql.language.FieldDefinition
 import org.neo4j.graphql.MetaProvider
 import org.neo4j.graphql.NodeFacade
 import org.neo4j.graphql.Translator
-import org.neo4j.graphql.handler.projection.ProjectionRepository
 
 class CreateTypeHandler(
         type: NodeFacade,
         fieldDefinition: FieldDefinition,
-        metaProvider: MetaProvider,
-        projectionRepository: ProjectionRepository
-) : BaseDataFetcher(type, fieldDefinition, metaProvider, projectionRepository) {
+        metaProvider: MetaProvider
+) : BaseDataFetcher(type, fieldDefinition, metaProvider) {
 
     override fun generateCypher(variable: String, field: Field, projectionProvider: () -> Translator.Cypher, ctx: Translator.Context): Translator.Cypher {
         val properties = properties(variable, field.arguments)
