+++ src/main/kotlin/org/neo4j/graphql/handler/DeleteHandler.kt
@@ -5,7 +5,6 @@ import graphql.language.FieldDefinition
 import org.neo4j.graphql.MetaProvider
 import org.neo4j.graphql.NodeFacade
 import org.neo4j.graphql.Translator
-import org.neo4j.graphql.handler.projection.ProjectionRepository
 import org.neo4j.graphql.quote
 
 class DeleteHandler(
@@ -13,9 +12,8 @@ class DeleteHandler(
         private val idField: FieldDefinition,
         fieldDefinition: FieldDefinition,
         metaProvider: MetaProvider,
-        projectionRepository: ProjectionRepository,
-        val isRealtion: Boolean = type.isRelationType()
-) : BaseDataFetcher(type, fieldDefinition, metaProvider, projectionRepository) {
+        private val isRealtion: Boolean = type.isRelationType()
+) : BaseDataFetcher(type, fieldDefinition, metaProvider) {
 
     override fun generateCypher(variable: String, field: Field, projectionProvider: () -> Translator.Cypher, ctx: Translator.Context): Translator.Cypher {
         val idArg = field.arguments.first { it.name == idField.name }
