--- src/main/kotlin/org/neo4j/graphql/handler/projection/ProjectionHandler.kt
@@ -1,163 +1,0 @@
-package org.neo4j.graphql.handler.projection
-
-import graphql.language.*
-import org.neo4j.graphql.*
-import org.neo4j.graphql.Translator.*
-
-
-class ProjectionHandler(metaProvider: MetaProvider) : ProjectionBase(metaProvider) {
-
-    fun projectFields(variable: String, field: Field, nodeType: NodeFacade, ctx: Context, variableSuffix: String?): Cypher {
-        val properties = field.selectionSet.selections.flatMap {
-            when (it) {
-                is Field -> listOf(projectField(variable, it, nodeType, ctx, variableSuffix))
-                is InlineFragment -> projectInlineFragment(variable, it, nodeType, ctx, variableSuffix)
-                is FragmentSpread -> projectNamedFragments(variable, it, nodeType, ctx, variableSuffix)
-                else -> emptyList()
-            }
-        }
-
-        val projection = properties.joinToString(", ", "{ ", " }") { it.query }
-        val params = properties.map { it.params }.fold(emptyMap<String, Any?>()) { res, map -> res + map }
-        return Cypher("$variable $projection", params)
-    }
-
-    private fun projectField(variable: String, field: Field, type: NodeFacade, ctx: Context, variableSuffix: String?): Cypher {
-        if (field.name == "__typename") {
-            val paramName = paramName(variable, "validTypes", null)
-            val validTypeLabels = metaProvider.getValidTypeLabels(type)
-            return Cypher("${field.aliasOrName()}: head( [ label in labels($variable) WHERE label in $$paramName ] )",
-                    mapOf(paramName to validTypeLabels))
-        }
-        val fieldDefinition = type.getFieldDefinition(field.name)
-                ?: throw IllegalStateException("No field ${field.name} in ${type.name()}")
-        val cypherDirective = fieldDefinition.cypherDirective()
-        val isObjectField = metaProvider.getNodeType(fieldDefinition.type.name()) != null
-        return cypherDirective?.let {
-            val directive = cypherDirective(variable, fieldDefinition, field, it, listOf(CypherArgument("this", "this", variable)))
-            if (isObjectField) {
-                val patternComprehensions = projectListComprehension(variable, field, fieldDefinition, ctx, directive, variableSuffix)
-                Cypher(field.aliasOrName() + ":" + patternComprehensions.query, patternComprehensions.params)
-            } else
-                Cypher(field.aliasOrName() + ":" + directive.query, directive.params)
-
-        } ?: when {
-            isObjectField -> {
-                val patternComprehensions = projectRelationship(variable, field, fieldDefinition, type, ctx, variableSuffix)
-                Cypher(field.aliasOrName() + ":" + patternComprehensions.query, patternComprehensions.params)
-            }
-            fieldDefinition.isNativeId() -> Cypher("${field.aliasOrName()}:ID($variable)")
-            else -> {
-                val dynamicPrefix = fieldDefinition.dynamicPrefix(metaProvider)
-                when {
-                    dynamicPrefix != null -> Cypher("${field.aliasOrName()}:apoc.map.fromPairs([key IN keys($variable) WHERE key STARTS WITH \"$dynamicPrefix\"| [substring(key,${dynamicPrefix.length}), $variable[key]]])")
-                    field.aliasOrName() == field.propertyName(fieldDefinition) -> Cypher("." + field.propertyName(fieldDefinition))
-                    else -> Cypher(field.aliasOrName() + ":" + variable + "." + field.propertyName(fieldDefinition))
-                }
-            }
-        }
-    }
-
-    private fun projectNamedFragments(variable: String, fragmentSpread: FragmentSpread, type: NodeFacade, ctx: Context, variableSuffix: String?) =
-            ctx.fragments.getValue(fragmentSpread.name).let {
-                projectFragment(it.typeCondition.name, type, variable, ctx, variableSuffix, it.selectionSet)
-            }
-
-    private fun projectFragment(fragmentTypeName: String?, type: NodeFacade, variable: String, ctx: Context, variableSuffix: String?, selectionSet: SelectionSet): List<Cypher> {
-        val fragmentType = metaProvider.getNodeType(fragmentTypeName)
-        return if (fragmentType == type) {
-            // these are the nested fields of the fragment
-            // it could be that we have to adapt the variable name too, and perhaps add some kind of rename
-            selectionSet.selections.filterIsInstance<Field>().map { projectField(variable, it, fragmentType, ctx, variableSuffix) }
-        } else {
-            emptyList()
-        }
-    }
-
-    private fun projectInlineFragment(variable: String, fragment: InlineFragment, type: NodeFacade, ctx: Context, variableSuffix: String?) =
-            projectFragment(fragment.typeCondition.name, type, variable, ctx, variableSuffix, fragment.selectionSet)
-
-
-    private fun projectRelationship(variable: String, field: Field, fieldDefinition: FieldDefinition, parent: NodeFacade, ctx: Context, variableSuffix: String?): Cypher {
-        return when (parent.getDirective(DirectiveConstants.RELATION) != null) {
-            true -> projectRelationshipParent(variable, field, fieldDefinition, ctx, variableSuffix)
-            else -> projectRichAndRegularRelationship(variable, field, fieldDefinition, parent, ctx)
-        }
-    }
-
-    private fun projectListComprehension(variable: String, field: Field, fieldDefinition: FieldDefinition, ctx: Context, expression: Cypher, variableSuffix: String?): Cypher {
-        val fieldObjectType = metaProvider.getNodeType(fieldDefinition.type.name()) ?: return Cypher.EMPTY
-        val fieldType = fieldDefinition.type
-        val childVariable = variable + field.name.capitalize()
-
-        // val where = where(childVariable, fieldDefinition, fieldObjectType, propertyArguments(field))
-        val fieldProjection = projectFields(childVariable, field, fieldObjectType, ctx, variableSuffix)
-
-        val comprehension = "[$childVariable IN ${expression.query} | ${fieldProjection.query}]"
-        val skipLimit = SkipLimit(childVariable, field.arguments)
-        val slice = skipLimit.slice(fieldType.isList())
-        return Cypher(comprehension + slice.query, (expression.params + fieldProjection.params + slice.params)) // + where.params
-
-    }
-
-    private fun relationshipInfoInCorrectDirection(fieldObjectType: NodeFacade, relInfo0: RelationshipInfo, parent: NodeFacade, relDirectiveField: RelationshipInfo?): RelationshipInfo {
-        val startField = fieldObjectType.getFieldDefinition(relInfo0.startField)!!
-        val endField = fieldObjectType.getFieldDefinition(relInfo0.endField)!!
-        val startFieldTypeName = startField.type.inner().name()
-        val inverse = startFieldTypeName != parent.name() || startField.type.name() == endField.type.name() && relDirectiveField?.out != relInfo0.out
-        return if (inverse) relInfo0.copy(out = relInfo0.out?.not(), startField = relInfo0.endField, endField = relInfo0.startField) else relInfo0
-    }
-
-    private fun projectRelationshipParent(variable: String, field: Field, fieldDefinition: FieldDefinition, ctx: Context, variableSuffix: String?): Cypher {
-        val fieldObjectType = metaProvider.getNodeType(fieldDefinition.type.name()) ?: return Cypher.EMPTY
-
-        val fieldProjection = projectFields(variable + (variableSuffix?.capitalize()
-                ?: ""), field, fieldObjectType, ctx, variableSuffix)
-
-        return Cypher(fieldProjection.query)
-    }
-
-    private fun projectRichAndRegularRelationship(variable: String, field: Field, fieldDefinition: FieldDefinition, parent: NodeFacade, ctx: Context): Cypher {
-        val fieldType = fieldDefinition.type
-        val fieldTypeName = fieldDefinition.type.name()!!
-        val nodeType = metaProvider.getNodeType(fieldTypeName)
-                ?: throw IllegalArgumentException("$fieldTypeName cannot be converted to a NodeGraphQlFacade")
-
-        // todo combine both nestings if rel-entity
-        val relDirectiveObject = nodeType.getDirective(DirectiveConstants.RELATION)?.let { relDetails(nodeType, it) }
-        val relDirectiveField = fieldDefinition.getDirective(DirectiveConstants.RELATION)?.let { relDetails(nodeType, it) }
-
-        val (relInfo0, isRelFromType) =
-                relDirectiveObject?.let { it to true }
-                        ?: relDirectiveField?.let { it to false }
-                        ?: throw IllegalStateException("Field $field needs an @relation directive")
-
-        val relInfo = if (isRelFromType) relationshipInfoInCorrectDirection(nodeType, relInfo0, parent, relDirectiveField) else relInfo0
-
-        val (inArrow, outArrow) = relInfo.arrows
-
-        val childVariable = variable + field.name.capitalize()
-
-        val (endNodePattern, variableSuffix) = when {
-            isRelFromType -> {
-                val label = nodeType.getFieldDefinition(relInfo.endField!!)!!.type.inner().name()
-                ("$childVariable${relInfo.endField.capitalize()}:$label" to relInfo.endField)
-            }
-            else -> ("$childVariable:${nodeType.name()}" to null)
-        }
-
-        val relPattern = if (isRelFromType) "$childVariable:${relInfo.relType}" else ":${relInfo.relType}"
-
-        // TODO abg cascade into type
-        val where = where(childVariable, fieldDefinition, nodeType, propertyArguments(field), ctx)
-        val fieldProjection = projectFields(childVariable, field, nodeType, ctx, variableSuffix)
-
-        val comprehension = "[($variable)$inArrow-[$relPattern]-$outArrow($endNodePattern)${where.query} | ${fieldProjection.query}]"
-        val skipLimit = SkipLimit(childVariable, field.arguments)
-        val slice = skipLimit.slice(fieldType.isList())
-        return Cypher(comprehension + slice.query, (where.params + fieldProjection.params + slice.params))
-    }
-
-    private fun relDetails(type: NodeFacade, relDirective: Directive) =
-            relDetails(type) { name, defaultValue -> metaProvider.getDirectiveArgument(relDirective, name, defaultValue) }
-}
